requirements.txt
pytest
odoo.tests.common
